--- src/components/timecapsule/main/Menu.tsx
@@ -1,13 +1,0 @@
-import * as S from '../../../styles/timecapsule/main/Menu.style';
-
-const Menu = () => {
-	return (
-		<S.MenuContainer>
-			<S.MenuItem>나의 Todo</S.MenuItem>
-			<S.MenuItem>나의 편지함</S.MenuItem>
-			<S.MenuItem>로그아웃</S.MenuItem>
-		</S.MenuContainer>
-	);
-};
-
-export default Menu;
