--- src/styles/todo/main/TodoMenu.style.ts
@@ -1,25 +1,0 @@
-import styled from 'styled-components';
-
-export const MenuContainer = styled.div`
-	display: flex;
-	flex-direction: column;
-	position: fixed;
-	top: 140px;
-	right: 65px;
-	gap: 30px;
-`;
-
-export const MenuItem = styled.button`
-	padding: 15px 20px;
-	color: black;
-	font-size: 18px;
-	font-weight: 600;
-	border: 2px solid lightgray;
-	border-radius: 50px;
-	cursor: pointer;
-
-	&:hover {
-		background-color: gray;
-		color: white;
-	}
-`;
